fix(playerbot): Move GuildEventCoordinator types to interface, remove duplicate destructors
- Moved GuildEventType, EventStatus enums to IGuildEventCoordinator.h
- Moved EventCoordinationProfile, EventParticipation, EventMetrics structs to interface
- Added copy constructor and assignment operator for EventMetrics (atomic members)
- Removed duplicate structs from GuildEventCoordinator.h
- Removed duplicate destructor from GuildBankManager.h

Next: Fix HybridAIController namespace mismatch error
